Extract discount computation into offer
